requirements.txt
supabase==2.3.0
python-jose[cryptography]==3.3.0
openpyxl==3.1.2
anthropic==0.39.0
